requirements.txt
# Retail Analytics Requirements
# Based on main project requirements with retail-specific additions

agent-framework-core>=1.0.0b251016
agent-framework-azure-ai>=1.0.0b251016
azure-ai-projects>=1.0.0b12
azure-identity>=1.23.1
python-dotenv>=1.1.1
pydantic>=2.11.7
fastapi>=0.116.0
uvicorn>=0.35.0
pyodbc>=5.2.0
asyncio
